backend/requirements.txt
# Jarvis AI Desktop Assistant - Backend
# Python 3.11+
#
# After cloning, install with:
#   pip install -r requirements.txt
#   playwright install chromium

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
python-multipart>=0.0.12
aiosqlite>=0.20.0
chromadb>=0.5.0
openai-whisper>=20240930
numpy>=1.26.0
soundfile>=0.12.1
pyttsx3>=2.90
pvporcupine>=3.0.2
sounddevice>=0.5.1
mss>=9.0.1
Pillow>=10.4.0
PyQt6>=6.7.1
playwright>=1.49.0
psutil>=6.1.0
google-api-python-client>=2.154.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
pyperclip>=1.9.0
tenacity>=9.0.0
boto3>=1.35.0
googlemaps>=4.10.0
pyautogui>=0.9.54
opencv-python>=4.8.0
mediapipe>=0.10.0

# -- Shopping & Learning skills -------------------------------------------------
# Install: pip install requests beautifulsoup4 yt-dlp
requests>=2.31.0
beautifulsoup4>=4.12.3
yt-dlp>=2024.1.0

# -- Food & Grocery MCP servers -------------------------------------------------
# Install: pip install playwright && playwright install chromium
playwright>=1.49.0
mcp>=1.2.0
fastmcp>=0.4.1

# -- Web Agent skill - browser automation --------------------------------------
# Install: pip install browser-use openai
# Then:    playwright install chromium
browser-use>=0.1.40
openai>=1.30.0
# Optional for lower-latency Windows capture:
# dxcam>=0.0.5
# Provides better product matching than the built-in scrapers.
# Install: pip install "git+https://github.com/laxmanclo/PriceHunter.git"
# Minimum deps for scraping only (no RAG/AI):
#   pip install aiohttp fake-useragent loguru fuzzywuzzy python-Levenshtein lxml
# Full install with AI insights (adds torch, faiss, sentence-transformers ~2GB):
#   pip install -r vendor/PriceHunter/requirements.txt
#   then set PRICEHUNTER_RAG=true in backend/.env
# git+https://github.com/laxmanclo/PriceHunter.git   uncomment to install
